UPD: move methods relative to aps, from naive to aps
