+++ mapie/conformity_scores/sets/aps.py
@@ -1,12 +1,16 @@
-from typing import Optional, cast
+from typing import Optional, Union, cast
 
 import numpy as np
 from sklearn.dummy import check_random_state
 
 from mapie.conformity_scores.sets.naive import Naive
-from mapie.conformity_scores.sets.utils import get_true_label_cumsum_proba
+from mapie.conformity_scores.sets.utils import (
+    check_include_last_label, check_proba_normalized,
+    get_true_label_cumsum_proba
+)
 from mapie.estimator.classifier import EnsembleClassifier
 
+from mapie._machine_precision import EPSILON
 from mapie._typing import NDArray
 from mapie.utils import compute_quantiles
 
@@ -37,6 +41,49 @@ class APS(Naive):
 
     def __init__(self) -> None:
         super().__init__()
+
+    def get_predictions(
+        self,
+        X: NDArray,
+        alpha_np: NDArray,
+        estimator: EnsembleClassifier,
+        agg_scores: Optional[str] = "mean",
+        **kwargs
+    ) -> NDArray:
+        """
+        Get predictions from an EnsembleClassifier.
+
+        Parameters:
+        -----------
+        X: NDArray of shape (n_samples, n_features)
+            Observed feature values.
+
+        alpha_np: NDArray of shape (n_alpha,)
+            NDArray of floats between ``0`` and ``1``, represents the
+            uncertainty of the confidence interval.
+
+        estimator: EnsembleClassifier
+            Estimator that is fitted to predict y from X.
+
+        agg_scores: Optional[str]
+            Method to aggregate the scores from the base estimators.
+            If "mean", the scores are averaged. If "crossval", the scores are
+            obtained from cross-validation.
+
+            By default ``"mean"``.
+
+        Returns:
+        --------
+        NDArray
+            Array of predictions.
+        """
+        y_pred_proba = estimator.predict(X, agg_scores)
+        y_pred_proba = check_proba_normalized(y_pred_proba, axis=1)
+        if agg_scores != "crossval":
+            y_pred_proba = np.repeat(
+                y_pred_proba[:, :, np.newaxis], len(alpha_np), axis=2
+            )
+        return y_pred_proba
 
     def get_conformity_scores(
         self,
@@ -124,3 +171,229 @@ class APS(Naive):
             quantiles_ = (n + 1) * (1 - alpha_np)
 
         return quantiles_
+
+    def _compute_vs_parameter(
+        self,
+        y_proba_last_cumsumed: NDArray,
+        threshold: NDArray,
+        y_pred_proba_last: NDArray,
+        prediction_sets: NDArray,
+        **kwargs
+    ) -> NDArray:
+        """
+        Compute the V parameters from Romano+(2020).
+
+        Parameters:
+        -----------
+        y_proba_last_cumsumed: NDArray of shape (n_samples, n_alpha)
+            Cumulated score of the last included label.
+
+        threshold: NDArray of shape (n_alpha,) or shape (n_samples_train,)
+            Threshold to compare with y_proba_last_cumsum.
+
+        y_pred_proba_last: NDArray of shape (n_samples, 1, n_alpha)
+            Last included probability.
+
+        predicition_sets: NDArray of shape (n_samples, n_alpha)
+            Prediction sets.
+
+        Returns:
+        --------
+        NDArray of shape (n_samples, n_alpha)
+            Vs parameters.
+        """
+        # compute V parameter from Romano+(2020)
+        vs = (
+            (y_proba_last_cumsumed - threshold.reshape(1, -1)) /
+            y_pred_proba_last[:, 0, :]
+        )
+        return vs
+
+    def _add_random_tie_breaking(
+        self,
+        prediction_sets: NDArray,
+        y_pred_index_last: NDArray,
+        y_pred_proba_cumsum: NDArray,
+        y_pred_proba_last: NDArray,
+        threshold: NDArray,
+        random_state: Optional[Union[int, np.random.RandomState]] = None,
+        **kwargs
+    ) -> NDArray:
+        """
+        Randomly remove last label from prediction set based on the
+        comparison between a random number and the difference between
+        cumulated score of the last included label and the quantile.
+
+        Parameters
+        ----------
+        prediction_sets: NDArray of shape
+            (n_samples, n_classes, n_threshold)
+            Prediction set for each observation and each alpha.
+
+        y_pred_index_last: NDArray of shape (n_samples, threshold)
+            Index of the last included label.
+
+        y_pred_proba_cumsum: NDArray of shape (n_samples, n_classes)
+            Cumsumed probability of the model in the original order.
+
+        y_pred_proba_last: NDArray of shape (n_samples, 1, threshold)
+            Last included probability.
+
+        threshold: NDArray of shape (n_alpha,) or shape (n_samples_train,)
+            Threshold to compare with y_proba_last_cumsum, can be either:
+
+            - the quantiles associated with alpha values when
+                ``cv`` == "prefit", ``cv`` == "split"
+                or ``agg_scores`` is "mean"
+
+            - the conformity score from training samples otherwise (i.e., when
+            ``cv`` is CV splitter and ``agg_scores`` is "crossval")
+
+        method: str
+            Method that determines how to remove last label in the prediction
+            set.
+
+            - if "cumulated_score" or "aps", compute V parameter
+                from Romano+(2020)
+
+            - else compute V parameter from Angelopoulos+(2020)
+
+        lambda_star: Optional[Union[NDArray, float]] of shape (n_alpha):
+            Optimal value of the regulizer lambda.
+
+        k_star: Optional[NDArray] of shape (n_alpha):
+            Optimal value of the regulizer k.
+
+        Returns
+        -------
+        NDArray of shape (n_samples, n_classes, n_alpha)
+            Updated version of prediction_sets with randomly removed labels.
+        """
+        # get cumsumed probabilities up to last retained label
+        y_proba_last_cumsumed = np.squeeze(
+            np.take_along_axis(
+                y_pred_proba_cumsum,
+                y_pred_index_last,
+                axis=1
+            ), axis=1
+        )
+
+        # get the V parameter from Romano+(2020) or Angelopoulos+(2020)
+        vs = self._compute_vs_parameter(
+            y_proba_last_cumsumed,
+            threshold,
+            y_pred_proba_last,
+            prediction_sets
+        )
+
+        # get random numbers for each observation and alpha value
+        random_state = check_random_state(random_state)
+        random_state = cast(np.random.RandomState, random_state)
+        us = random_state.uniform(size=(prediction_sets.shape[0], 1))
+        # remove last label from comparison between uniform number and V
+        vs_less_than_us = np.less_equal(vs - us, EPSILON)
+        np.put_along_axis(
+            prediction_sets,
+            y_pred_index_last,
+            vs_less_than_us[:, np.newaxis, :],
+            axis=1
+        )
+        return prediction_sets
+
+    def get_prediction_sets(
+        self,
+        y_pred_proba: NDArray,
+        conformity_scores: NDArray,
+        alpha_np: NDArray,
+        estimator: EnsembleClassifier,
+        agg_scores: Optional[str] = "mean",
+        include_last_label: Optional[Union[bool, str]] = True,
+        **kwargs
+    ) -> NDArray:
+        """
+        Generate prediction sets based on the probability predictions,
+        the conformity scores and the uncertainty level.
+
+        Parameters:
+        -----------
+        y_pred_proba: NDArray of shape (n_samples, n_classes)
+            Target prediction.
+
+        conformity_scores: NDArray of shape (n_samples,)
+            Conformity scores for each sample.
+
+        alpha_np: NDArray of shape (n_alpha,)
+            NDArray of floats between 0 and 1, representing the uncertainty
+            of the confidence interval.
+
+        estimator: EnsembleClassifier
+            Estimator that is fitted to predict y from X.
+
+        agg_scores: Optional[str]
+            Method to aggregate the scores from the base estimators.
+            If "mean", the scores are averaged. If "crossval", the scores are
+            obtained from cross-validation.
+
+            By default ``"mean"``.
+
+        include_last_label: Optional[Union[bool, str]]
+            Whether or not to include last label in prediction sets.
+            Choose among ``False``, ``True``  or ``"randomized"``.
+
+            By default, ``True``.
+
+        Returns:
+        --------
+        NDArray
+            Array of quantiles with respect to alpha_np.
+        """
+        include_last_label = check_include_last_label(include_last_label)
+
+        # specify which thresholds will be used
+        if estimator.cv == "prefit" or agg_scores in ["mean"]:
+            thresholds = self.quantiles_
+        else:
+            thresholds = conformity_scores.ravel()
+
+        # sort labels by decreasing probability
+        y_pred_proba_cumsum, y_pred_index_last, y_pred_proba_last = (
+            self._get_last_included_proba(
+                y_pred_proba,
+                thresholds,
+                include_last_label,
+                prediction_phase=True,
+                **kwargs
+            )
+        )
+        # get the prediction set by taking all probabilities above the last one
+        if estimator.cv == "prefit" or agg_scores in ["mean"]:
+            y_pred_included = np.greater_equal(
+                y_pred_proba - y_pred_proba_last, -EPSILON
+            )
+        else:
+            y_pred_included = np.less_equal(
+                y_pred_proba - y_pred_proba_last, EPSILON
+            )
+        # remove last label randomly
+        if include_last_label == "randomized":
+            y_pred_included = self._add_random_tie_breaking(
+                y_pred_included,
+                y_pred_index_last,
+                y_pred_proba_cumsum,
+                y_pred_proba_last,
+                thresholds,
+                self.random_state,
+                **kwargs
+            )
+        if estimator.cv == "prefit" or agg_scores in ["mean"]:
+            prediction_sets = y_pred_included
+        else:
+            # compute the number of times the inequality is verified
+            prediction_sets_summed = y_pred_included.sum(axis=2)
+            prediction_sets = np.less_equal(
+                prediction_sets_summed[:, :, np.newaxis]
+                - self.quantiles_[np.newaxis, np.newaxis, :],
+                EPSILON
+            )
+
+        return prediction_sets
